+++ sg-api/src/main/java/com/github/tangyi/api/user/service/IQiNiuService.java
@@ -27,7 +27,7 @@ public interface IQiNiuService {
 
 	String getDownloadUrl(AttachGroup group, String attachName);
 
-	String getDownloadUrl(String fileName, int expire);
+	String getDownloadUrl(String fileName, long expire);
 
 	boolean isNotDefaultGroup(Attachment attachment);
